pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tcc-sentiment-analysis"
version = "0.1.0"
description = "Análise de sentimento em notícias financeiras para previsão de tendências de mercado"
authors = [{ name = "Marcelo", email = "marcelo.santos.c@grad.ufsc.br" }]
requires-python = ">=3.12"
dependencies = [
    "requests",
    "python-dotenv",
    "transformers",
    "yfinance",
    "torch",
    "pandas",
    "tqdm",
    "pytorch-forecasting",
    "pyarrow",
    "pandas-ta==0.4.71b0",
    "optuna",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::UserWarning",
]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.ruff]
target-version = "py312"
line-length = 88
# Note: 'fix = true' is deprecated in config; use 'ruff check --fix' in terminal instead

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]
ignore = ["E501"]
fixable = ["I", "UP", "F"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
lines-between-types = 1

[tool.coverage.run]
source = ["src"]
omit = [
    "src/main.py",
    "src/adapters/*",
]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == \"__main__\":",
    "raise AssertionError",
    "raise NotImplementedError",
]
precision = 2
show_missing = true
